Move MultivarJson to protocol
