+++ src/ripple/app/ledger/BookListeners.h
@@ -20,7 +20,6 @@
 #ifndef RIPPLE_APP_LEDGER_BOOKLISTENERS_H_INCLUDED
 #define RIPPLE_APP_LEDGER_BOOKLISTENERS_H_INCLUDED
 
-#include <ripple/json/MultivarJson.h>
 #include <ripple/net/InfoSub.h>
 #include <ripple/protocol/ApiVersion.h>
