+++ src/ripple/app/ledger/OrderBookDB.h
@@ -23,7 +23,6 @@
 #include <ripple/app/ledger/AcceptedLedgerTx.h>
 #include <ripple/app/ledger/BookListeners.h>
 #include <ripple/app/main/Application.h>
-#include <ripple/json/MultivarJson.h>
 #include <ripple/protocol/ApiVersion.h>
 
 #include <mutex>
